re-work main view component
