+++ src/components/Cards.jsx
@@ -2,19 +2,17 @@ import "../assets/css/Cards.css";
 
 const Cards = (props) => {
   return (
+    // filter cards by list id. render each card
     <div className="cards">
-      {/* iterate all cards on board */}
-      {props.currentCards.filter(e => e.idList === props.listId).map((e, i) => {
+      {props.currentCards
+        .filter((e) => e.idList === props.listId)
+        .map((e, i) => {
           return (
             <div className="cards__card" key={i}>
-              <div className="cards__name">
-              {e.name}
-              </div>
+              <div className="cards__name">{e.name}</div>
             </div>
-
           );
-        } 
-      )}
+        })}
     </div>
   );
 };
